+++ Classes/Data/MLDB.lua
@@ -152,7 +152,6 @@ function private:BuildMLDB()
         rejectTrade = db.rejectTrade or nil,
         outOfRaid = db.outOfRaid or nil,
         autoGroupLoot = db.autoGroupLoot or nil,
-		lootDecor = db.autoGroupLootDecor or nil,
     }
     self.isBuilt = true
     return self.mldb