+++ Classes/Utils/GroupLoot.lua
@@ -35,13 +35,6 @@ GroupLoot.Predicates = {
 	qualityUpperBound = function (_, _, _, quality)
 		return not quality or quality < Enum.ItemQuality.Legendary
 	end,
-	--- Decor items are toggleable
-	decor = function(rollID, itemID, status)
-		local itemClass = select(6, C_Item.GetItemInfoInstant(itemID))
-		return itemClass ~= Enum.ItemClass.Housing or
-			itemClass == Enum.ItemClass.Housing and addon.mldb and
-			addon.mldb.lootDecor
-	end,
 }
 
 function GroupLoot:OnInitialize()